+++ src/cloudflare-ddns/Program.cs
@@ -3,7 +3,6 @@ using CloudflareDDNS.Service;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Serilog;
-using Serilog.Formatting.Compact;
 using Mono.Options;
 using System.Reflection;
 using System;
@@ -19,10 +18,11 @@ namespace CloudflareDDNS
             var configFile = "config.json";
             var showHelp = false;
             var showVersion = false;
-            var verbose = false;
+            var logLevel = "info";
             var options = new OptionSet
             {
-                { "c|config", "Path to the config.json file, defaults to the current working directory.", (string c) => configFile = c},
+                { "c|config", "Path to the config.json file, defaults to the current working directory.", (string c) => configFile = c },
+                { "l|log-level=", "Set the log level (verbose, info, warning, error), defaults to info", (string l) => logLevel = l ?? logLevel},
                 { "v|version", "Show current version.", v => showVersion = v != null },
                 { "?|h|help", "Show help information", h => showHelp = h != null },
             };
@@ -50,7 +50,7 @@ namespace CloudflareDDNS
                     return 0;
                 }
 
-                await Execute(configFile, verbose);
+                await Execute(configFile, logLevel);
                 return 0;
             }
             catch (OptionException ex)
@@ -70,7 +70,7 @@ namespace CloudflareDDNS
             }
         }
 
-        static async Task Execute(string configFile, bool verbose)
+        static async Task Execute(string configFile, string logLevel)
         {
             var appConfig = await Config.LoadFromAsync(configFile);
 
@@ -78,11 +78,11 @@ namespace CloudflareDDNS
             var fileSystem = new FileSystem();
             var logger = new LoggerConfiguration()
                 .WriteTo.Console()
-                .MinimumLevel.Is(verbose ? LogEventLevel.Verbose : LogEventLevel.Information)
+                .MinimumLevel.Is(MapToLogEventLevel(logLevel))
                 .CreateLogger();
 
             var ipResolver = new IPResolverService(client, logger, new IPResolverServiceOptions(appConfig.IPv4Resolver, appConfig.IPv6Resolver));
-            var cloudflare = new CloudflareApi(client, new CloudflareApiOptions(appConfig.ApiKey));
+            var cloudflare = new CloudflareApi(client, new CloudflareApiOptions(appConfig.ApiToken));
             var ddns = new DynamicDnsService(logger, cloudflare);
 
             await SyncDDNSCommand(logger, ipResolver, ddns, appConfig);
@@ -96,6 +96,16 @@ namespace CloudflareDDNS
                 await ddns.UpsertRecords(item.ZoneId, item.Domain, item.Proxied, ips);
             }
         }
+
+        public static LogEventLevel MapToLogEventLevel(string logLevel) =>
+            logLevel switch
+            {
+                "verbose" => LogEventLevel.Verbose,
+                "info" => LogEventLevel.Information,
+                "warning" => LogEventLevel.Warning,
+                "error" => LogEventLevel.Error,
+                _ => throw new ArgumentException($"Invalid log level provided ({logLevel}), only verbose, info, warning, error are supported!")
+            };
 
         private static string GetVersion()
             => typeof(Program).Assembly.GetCustomAttribute<AssemblyInformationalVersionAttribute>()?.InformationalVersion ?? "0.0.0";