+++ test/CloudflareDDNS.Tests/ConfigTests.cs
@@ -12,8 +12,13 @@ namespace CloudflareDDNS.Tests
     {
         private readonly JsonSerializerOptions serializerOptions = new JsonSerializerOptions
         {
-            IgnoreNullValues = false
+            IgnoreNullValues = false,
         };
+
+        public ConfigTests()
+        {
+            Environment.SetEnvironmentVariable(Config.ENV_CLOUDFLARE_API_TOKEN, null);
+        }
 
         [Fact]
         public async Task Config_LoadFrom_ParsesAMinimalValidConfigFile()
@@ -22,7 +27,7 @@ namespace CloudflareDDNS.Tests
             var contents = JsonSerializer.Serialize(
                 new
                 {
-                    apiKey = "123",
+                    apiToken = "123",
                     dns = new[]
                     {
                         new { zoneId = "456", domain = "test" }
@@ -41,7 +46,7 @@ namespace CloudflareDDNS.Tests
             // Then
             Assert.Equal("https://api.ipify.org", config.IPv4Resolver);
             Assert.Equal("https://api6.ipify.org", config.IPv6Resolver);
-            Assert.Equal("123", config.ApiKey);
+            Assert.Equal("123", config.ApiToken);
             Assert.Single(config.Dns);
             Assert.Collection(
                 config.Dns,
@@ -63,7 +68,7 @@ namespace CloudflareDDNS.Tests
                 {
                     ipv4Resolver = "https://custom-ipv4.org",
                     ipv6Resolver = "https://custom-ipv6.org",
-                    apiKey = "123",
+                    apiToken = "123",
                     dns = new[]
                     {
                         new { zoneId = "456", domain = "test" }
@@ -82,6 +87,34 @@ namespace CloudflareDDNS.Tests
             // Then
             Assert.Equal("https://custom-ipv4.org", config.IPv4Resolver);
             Assert.Equal("https://custom-ipv6.org", config.IPv6Resolver);
+        }
+
+        [Fact]
+        public async Task Config_LoadFrom_UsesApiTokenFromEnvironmentVariableWhenProvided()
+        {
+            // 
+            Environment.SetEnvironmentVariable(Config.ENV_CLOUDFLARE_API_TOKEN, "8910");
+            var contents = JsonSerializer.Serialize(
+                new
+                {
+                    apiToken = "123",
+                    dns = new[]
+                    {
+                        new { zoneId = "456", domain = "test" }
+                    }
+                },
+                serializerOptions
+            );
+            var fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>
+            {
+                { @"~/config.json", new MockFileData(contents) },
+            });
+
+            // When
+            var config = await Config.LoadFromAsync("~/config.json", fileSystem);
+
+            // Then
+            Assert.Equal("8910", config.ApiToken);
         }
 
         [Fact]
@@ -107,8 +140,8 @@ namespace CloudflareDDNS.Tests
             var exception = await Assert.ThrowsAsync<InvalidConfigException>(() => Config.LoadFromAsync("~/config.json", fileSystem));
 
             // Then
-            Assert.Equal("ApiKey", exception.Property);
-            Assert.Equal("Property can't be null or empty!", exception.Message);
+            Assert.Equal("ApiToken", exception.Property);
+            Assert.Equal("Property 'ApiToken' can't be null or empty!", exception.Message);
         }
 
         [Theory]
@@ -120,7 +153,7 @@ namespace CloudflareDDNS.Tests
             var contents = JsonSerializer.Serialize(
                 new
                 {
-                    apiKey = "123",
+                    apiToken = "123",
                     dns = new[]
                     {
                         new { zoneId = zoneId, domain = domain }
@@ -138,8 +171,7 @@ namespace CloudflareDDNS.Tests
 
             // Then
             Assert.Equal(property, exception.Property);
-            Assert.Equal("Property can't be null or empty!", exception.Message);
+            Assert.Equal($"Property '{property}' can't be null or empty!", exception.Message);
         }
-
     }
 }